+++ Scaffer.h
@@ -38,7 +38,7 @@ class Scaffer
 		~Scaffer();
 
 		void read_info(char *, int);
-		int read_mappings(char *, char *, bool);
+		long int read_mappings(char *, char *, bool);
 		int write_scaffolds(char *, char *, long int);
 
 	private:
